+++ src/main/java/com/calendly/mini/exception/ErrorDTO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class Error {
+public class ErrorDTO {
 
     private String message ;
     private String details ;
